Initialize ata before graphics
ATA init is the long pole in the boot process, and its asynchronous.
move the graphics init after it so that ata and graphics initialize
in parallel

UFS is also a part of ATA and is also somewhat slow in probing.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
(cherry picked from commit d542604e87d73df8190578153b0950185e75ef69)